refactor: Change test file structure
