+++ src/test/presentational/Input.test.jsx
@@ -1,6 +1,6 @@
 import { fireEvent, render } from '@testing-library/react';
 
-import Input from '../presentational/Input';
+import Input from '../../presentational/Input';
 
 describe('<Input />', () => {
   const handleChange = jest.fn();
